+++ README.md
@@ -67,9 +67,10 @@
   - [x] (프로덕트 코드) `li` 태그 전부 삭제 -> 테스트 실패
   - [x] (테스트 실패를 성공시키기 위해) js를 이용해 동적으로 `li` 태그 하나만 제작 -> 테스트 성공
   - [x] `li` 태그를 컴포넌트로 제작
-  - [ ] `li` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
-    - [ ] 동적 인자 전달 -> 정적 텍스트를 어떻게 관리할거냐 -> 입력한 값도 추가가 되어야하고, 기존에 있는 값도 추가가 되어 있어야한다. -> 도메인 (식당 정보들을 관리하고 있는 것! -> 인스턴스)
-    - [ ] `li` 테스트 코드 수정
+  - [x] `li` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
+    - [x] 동적 인자 전달 -> 정적 텍스트를 어떻게 관리할거냐 -> 입력한 값도 추가가 되어야하고, 기존에 있는 값도 추가가 되어 있어야한다.
+    - [ ] 도메인 (식당 정보들을 관리하고 있는 것! -> 인스턴스)
+    - [x] `li` 테스트 코드 수정
 
 - [ ] `list` (.restaurant-list)
   - [ ] (템플릿 코드 기반) `ul`(.restaurant-list) 태그를 바탕으로 테스트 작성
